requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
joblib==1.4.2
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.2
pdfminer.six==20231228
pytest==8.3.3
pytesseract==0.3.13
Pillow==10.4.0
rapidocr-onnxruntime==1.3.19
jaconv==0.3.4
python-multipart==0.0.9
